Tehtävä 4.21 tehty

+++ controllers/blogs.js
@@ -3,8 +3,6 @@ const jwt = require('jsonwebtoken')
 const Blog = require('../models/blog')
 const User = require('../models/user')
 
-
-
 const getTokenFrom = request => {
   const authorization = request.get('authorization')
   if (authorization && authorization.toLowerCase().startsWith('bearer ')) {
@@ -13,33 +11,37 @@ const getTokenFrom = request => {
   return null
 }
 
-blogsRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 })
-  response.json(blogs.map(blog => blog.toJSON()))
+blogsRouter.get('/', async (request, response, next) => {
+  try {
+    const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 })
+    response.json(blogs.map(blog => blog.toJSON()))
+  } catch(exception) {
+    next(exception)
+  }
 })
 
-blogsRouter.get('/:id', async (request, response) => {
-  const blogs = await Blog.findById(request.params.id)
-  response.json(blogs)
+blogsRouter.get('/:id', async (request, response, next) => {
+  try {
+    const blogs = await Blog.findById(request.params.id)
+    response.json(blogs)
+  } catch(exception) {
+    next(exception)
+  }
 })
-
-
 
 blogsRouter.post('/', async (request, response, next) => {
   const blog = new Blog(request.body)
-
   const token = getTokenFrom(request)
-  const decodedToken = jwt.verify(token, process.env.SECRET)
-  if (!token || !decodedToken.id) {
-    return response.status(401).json({ error: 'token missing or invalid' })
-  }
-  const user = await User.findById(decodedToken.id)
-
-  if (!blog.likes) {
-    blog.likes = 0
-  }
-  blog.user = user._id
   try {
+    const decodedToken = jwt.verify(token, process.env.SECRET)
+    if (!token || !decodedToken.id) {
+      return response.status(401).json({ error: 'token missing or invalid' })
+    }
+    const user = await User.findById(decodedToken.id)
+    if (!blog.likes) {
+      blog.likes = 0
+    }
+    blog.user = user._id
     const savedBlog = await blog.save()
     user.blogs = user.blogs.concat(savedBlog._id)
     await user.save()
@@ -49,12 +51,20 @@ blogsRouter.post('/', async (request, response, next) => {
   }
 })
 
-
-
 blogsRouter.delete('/:id', async (request, response, next) => {
+  const token = getTokenFrom(request)
   try {
-    await Blog.findByIdAndRemove(request.params.id)
-    response.status(204).end()
+    const decodedToken = jwt.verify(token, process.env.SECRET)
+    if (!token || !decodedToken.id) {
+      return response.status(401).json({ error: 'token missing or invalid' })
+    }
+    const blog = await Blog.findById(request.params.id)
+    if (blog.user.toString() === decodedToken.id.toString() ) {
+      await Blog.findByIdAndRemove(request.params.id)
+      response.status(204).end()
+    } else {
+      return response.status(401).json({ error: 'token missing or invalid' })
+    }
   } catch(exception) {
     next(exception)
   }
@@ -74,8 +84,5 @@ blogsRouter.put('/:id', async (request, response, next) => {
     next(exception)
   }
 })
-
-
-
 
 module.exports = blogsRouter
